Squelch compilation warnings (#517)
* Fix missing override warnings

* Fix warnings about signed vs unsigned comparison

* Fix misc compiler warnings

* Update use of deprecated functions

* Squelch unnecessary warnings

* Remove use of QTime un UpdateManager

* QTime.elapsed is deprecated. Instead of replacing it with QElapsedTimer, the UpdateManger is modified to use QTimer directly. We can expect the behaviour to be the same as before in most if not all scenarios. The new version is less resource intensive at the cost of possibly lower timing accuracy (although the previous version of the code also saw Ibis' tick period vary as a function of workload).

+++ IbisLib/triplecutplaneobject.h
@@ -81,7 +81,7 @@ public:
      */
     ///@{
     /** Get the number of images. */
-    int GetNumberOfImages() { return Images.size(); }
+    size_t GetNumberOfImages() { return Images.size(); }
     /** Get an ImageObject knowing its index. */
     ImageObject * GetImage( int index );
     /** Add an ImageObject knowing its ID. */
